+++ packages/react-native-audio-api/common/cpp/audioapi/core/utils/AudioStretcher.cpp
@@ -1,5 +1,6 @@
 #include <audioapi/core/sources/AudioBuffer.h>
 #include <audioapi/core/utils/AudioStretcher.h>
+#include <audioapi/core/utils/Constants.h>
 #include <audioapi/libs/audio-stretch/stretch.h>
 #include <audioapi/utils/AudioArray.h>
 #include <audioapi/utils/AudioBus.h>
@@ -7,8 +8,7 @@
 
 namespace audioapi {
 
-std::vector<int16_t> AudioStretcher::castToInt16Buffer(AudioBuffer &buffer)
-{
+std::vector<int16_t> AudioStretcher::castToInt16Buffer(AudioBuffer &buffer) {
   const size_t numChannels = buffer.getNumberOfChannels();
   const size_t numFrames = buffer.getLength();
 
@@ -24,8 +24,9 @@ std::vector<int16_t> AudioStretcher::castToInt16Buffer(AudioBuffer &buffer)
   return int16Buffer;
 }
 
-std::shared_ptr<AudioBuffer> AudioStretcher::changePlaybackSpeed(AudioBuffer buffer, float playbackSpeed)
-{
+std::shared_ptr<AudioBuffer> AudioStretcher::changePlaybackSpeed(
+    AudioBuffer buffer,
+    float playbackSpeed) {
   const float sampleRate = buffer.getSampleRate();
   const size_t outputChannels = buffer.getNumberOfChannels();
   const size_t numFrames = buffer.getLength();
@@ -36,20 +37,30 @@ std::shared_ptr<AudioBuffer> AudioStretcher::changePlaybackSpeed(AudioBuffer buf
 
   std::vector<int16_t> int16Buffer = castToInt16Buffer(buffer);
 
-  auto stretcher =
-      stretch_init(static_cast<int>(sampleRate / 333.0f), static_cast<int>(sampleRate / 55.0f), outputChannels, 0x1);
+  auto stretcher = stretch_init(
+      static_cast<int>(sampleRate / UPPER_FREQUENCY_LIMIT_DETECTION),
+      static_cast<int>(sampleRate / LOWER_FREQUENCY_LIMIT_DETECTION),
+      outputChannels,
+      0x1);
 
-  int maxOutputFrames = stretch_output_capacity(stretcher, static_cast<int>(numFrames), 1 / playbackSpeed);
+  int maxOutputFrames = stretch_output_capacity(
+      stretcher, static_cast<int>(numFrames), 1 / playbackSpeed);
   std::vector<int16_t> stretchedBuffer(maxOutputFrames * outputChannels);
 
   int outputFrames = stretch_samples(
-      stretcher, int16Buffer.data(), static_cast<int>(numFrames), stretchedBuffer.data(), 1 / playbackSpeed);
+      stretcher,
+      int16Buffer.data(),
+      static_cast<int>(numFrames),
+      stretchedBuffer.data(),
+      1 / playbackSpeed);
 
-  outputFrames += stretch_flush(stretcher, stretchedBuffer.data() + (outputFrames));
+  outputFrames +=
+      stretch_flush(stretcher, stretchedBuffer.data() + (outputFrames));
   stretchedBuffer.resize(outputFrames * outputChannels);
   stretch_deinit(stretcher);
 
-  auto audioBus = std::make_shared<AudioBus>(outputFrames, outputChannels, sampleRate);
+  auto audioBus =
+      std::make_shared<AudioBus>(outputFrames, outputChannels, sampleRate);
 
   for (int ch = 0; ch < outputChannels; ++ch) {
     auto channelData = audioBus->getChannel(ch)->getData();
