+++ packages/react-native-audio-api/common/cpp/audioapi/core/utils/AudioStretcher.h
@@ -20,10 +20,10 @@ class AudioStretcher {
   static std::vector<int16_t> castToInt16Buffer(AudioBuffer &buffer);
 
   [[nodiscard]] static inline int16_t floatToInt16(float sample) {
-    return static_cast<int16_t>(sample * 32768.0f);
+    return static_cast<int16_t>(sample * INT16_MAX);
   }
   [[nodiscard]] static inline float int16ToFloat(int16_t sample) {
-    return static_cast<float>(sample) / 32768.0f;
+    return static_cast<float>(sample) / INT16_MAX;
   }
 };
